UI 구현 & 버그픽스
1. HpBar 구현
2. DashCoolTimeBar 구현
3. 대쉬 조작 변경 - 우클릭시 마우스방향 이동
4. 대쉬로 벽뚫기 픽스

+++ Assets/Scripts/PlayerStat.cs
@@ -4,8 +4,8 @@ using UnityEngine;
 
 public class PlayerStat : MonoBehaviour
 {
-    public int currentHp;
-    public int maxHp = 100;
+    public float currentHp;
+    public float maxHp = 100f;
     public bool isDead;
 
     public int gold = 0;
